Refactor: Remove explicit source-level dependency between helper and Transformations
