+++ lib/inline_svg.rb
@@ -1,7 +1,9 @@
 require "inline_svg/version"
 require "inline_svg/action_view/helpers"
-require "inline_svg/finds_asset_paths"
 require "inline_svg/asset_file"
+require "inline_svg/finds_asset_paths"
+require "inline_svg/transform_pipeline"
+
 require "inline_svg/railtie" if defined?(Rails)
 require 'active_support/core_ext'
 require 'nokogiri'
